XW-3939 | intitializer/browser folder structure is no longer supported
