+++ app/initializers/ember-hammer.js
@@ -1,0 +1,22 @@
+/**
+ * Initializer for Ember Hammer
+ *
+ * @returns {void}
+ */
+export function initialize() {
+	if (typeof FastBoot === 'undefined') {
+		window.emberHammerOptions = {
+			hammerOptions: {
+				// we are using fastclick as this is advised by ember-hammer lib
+				ignoreEvents: [],
+				swipe_velocity: 0.1,
+				pan_threshold: 1
+			}
+		};
+	}
+}
+
+export default {
+	name: 'ember-hammer',
+	initialize
+};
