--- app/instance-initializers/browser/visit-source.js
@@ -1,17 +1,0 @@
-import config from '../../config/environment';
-
-/**
- * @returns {void}
- */
-export function initialize() {
-	if (typeof VisitSource === 'function') {
-		(new VisitSource('WikiaSessionSource', config.cookieDomain)).checkAndStore();
-		(new VisitSource('WikiaLifetimeSource', config.cookieDomain, false)).checkAndStore();
-	}
-}
-
-export default {
-	after: 'config',
-	name: 'visit-source',
-	initialize
-};
